re implemented cloud function in auth.ts and added functions export in config.ts
