+++ functions/tsconfig.json
@@ -11,5 +11,5 @@
   "compileOnSave": true,
   "include": [
     "src"
-  ]
+, "../backend/CloudFunctions.ts"  ]
 }
