+++ src/config.ts
@@ -1,7 +1,13 @@
 import { getAuth, connectAuthEmulator } from "firebase/auth";
 import { getFirestore, connectFirestoreEmulator } from "firebase/firestore";
+import { getApp } from "firebase/app";
+import { getFunctions, connectFunctionsEmulator } from "firebase/functions";
 
-const auth = getAuth();
-const db = getFirestore();
+export const auth = getAuth();
+export const db = getFirestore();
 connectAuthEmulator(auth, "http://127.0.0.1:9099");
 connectFirestoreEmulator(db, "127.0.0.1", 8080);
+
+
+export const functions = getFunctions(getApp());
+connectFunctionsEmulator(functions, "127.0.0.1", 5001);
